requirements.txt
streamlit
numpy
pillow
matplotlib
tensorflow
keras
